backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
supabase==2.21.1
markitdown[all]==0.1.2
google-generativeai==0.8.3
python-dotenv==1.0.1
jinja2==3.1.4
